Add files via upload
Tic-tac-toe is a fun game played by two players. Before starting the game, a 3x3 
square grid is formed using two vertical and two horizontal lines. The player who succeeds in 
placing three of their marks in a horizontal, vertical, or diagonal row is the winner. In this study, 
we develop a computer program where a player can challenge artificial intelligent heuristic 
approach (opponent) to win the game. We want to prove that artificial intelligent heuristic 
approach enables in giving a great performance for development of tic-tac-toe game.